データ成型処理を別ファイルに移動
